Make the rest of flatteners functional in anticipation of #381
