+++ lib/parse.js
@@ -1,181 +1,7 @@
 'use strict';
 
 var doctrine = require('doctrine');
-
-var flatteners = {
-  'abstract': flattenBoolean,
-  'access': function (result, tag) {
-    result.access = tag.access;
-  },
-  'alias': flattenName,
-  'arg': synonym('param'),
-  'argument': synonym('param'),
-  'augments': collect('augments'),
-  'author': flattenDescription,
-  'borrows': todo,
-  'callback': flattenDescription,
-  'class': flattenTypedName,
-  'classdesc': flattenDescription,
-  'const': synonym('constant'),
-  'constant': flattenTypedName,
-  'constructor': synonym('class'),
-  'constructs': todo,
-  'copyright': flattenDescription,
-  'default': todo,
-  'defaultvalue': synonym('default'),
-  'deprecated': flattenDescription,
-  'desc': synonym('description'),
-  'description': flattenDescription,
-  'emits': synonym('fires'),
-  'enum': todo,
-  'event': flattenDescription,
-  'example': function (result, tag) {
-    if (!tag.description) {
-      result.errors.push({
-        message: '@example without code',
-        commentLineNumber: tag.lineNumber
-      });
-      return;
-    }
-
-    if (!result.examples) {
-      result.examples = [];
-    }
-
-    var example = {
-      description: tag.description
-    };
-
-    if (tag.caption) {
-      example.caption = tag.caption;
-    }
-
-    result.examples.push(example);
-  },
-  'exception': synonym('throws'),
-  'exports': todo,
-  'extends': synonym('augments'),
-  'external': flattenDescription,
-  'file': flattenDescription,
-  'fileoverview': synonym('file'),
-  'fires': todo,
-  'func': synonym('function'),
-  'function': flattenName,
-  'global': function (result) {
-    result.scope = 'global';
-  },
-  'host': synonym('external'),
-  'ignore': flattenBoolean,
-  'implements': todo,
-  'inheritdoc': todo,
-  'inner': function (result) {
-    result.scope = 'inner';
-  },
-  'instance': function (result) {
-    result.scope = 'instance';
-  },
-  'interface': function (result, tag) {
-    result.interface = true;
-    if (tag.description) {
-      result.name = tag.description;
-    }
-  },
-  'kind': function (result, tag) {
-    result.kind = tag.kind;
-  },
-  'lends': flattenDescription,
-  'license': flattenDescription,
-  'linkcode': synonym('link'),
-  'linkplain': synonym('link'),
-  'listens': todo,
-  'member': flattenTypedName,
-  'memberof': flattenDescription,
-  'method': synonym('function'),
-  'mixes': todo,
-  'mixin': flattenName,
-  'module': flattenTypedName,
-  'name': flattenName,
-  'namespace': flattenTypedName,
-  'override': flattenBoolean,
-  'overview': synonym('file'),
-  'param': collect('params'),
-  'private': function (result) {
-    result.access = 'private';
-  },
-  'prop': synonym('property'),
-  'property': collect('properties'),
-  'protected': function (result) {
-    result.access = 'protected';
-  },
-  'public': function (result) {
-    result.access = 'public';
-  },
-  'readonly': flattenBoolean,
-  'requires': todo,
-  'return': synonym('returns'),
-  'returns': collect('returns'),
-  'see': collect('sees', true),
-  'since': flattenDescription,
-  'static': function (result) {
-    result.scope = 'static';
-  },
-  'summary': flattenDescription,
-  'this': todo,
-  'throws': collect('throws'),
-  'todo': collect('todos', true),
-  'tutorial': todo,
-  'type': todo,
-  'typedef': flattenTypedName,
-  'var': synonym('member'),
-  'variation': function (result, tag) {
-    result.variation = tag.variation;
-  },
-  'version': flattenDescription,
-  'virtual': synonym('abstract')
-};
-
-function todo() {}
-
-function collect(key, description) {
-  return function (result, tag) {
-    if (!result[key]) {
-      result[key] = [];
-    }
-    if (description) {
-      result[key].push(tag.description);
-    } else {
-      result[key].push(tag);
-    }
-  };
-}
-
-function synonym(key) {
-  return function (result, tag) {
-    return flatteners[key](result, tag, key);
-  };
-}
-
-function flattenBoolean(result, tag, key) {
-  result[key] = true;
-}
-
-function flattenName(result, tag, key) {
-  result[key] = tag.name;
-}
-
-function flattenDescription(result, tag, key) {
-  result[key] = tag.description;
-}
-
-function flattenTypedName(result, tag, key) {
-  result[key] = {
-    name: tag.name
-  };
-
-  if (tag.type) {
-    result[key].type = tag.type;
-  }
-}
+var flatteners = require('./flatteners');
 
 /**
  * Parse a comment with doctrine, decorate the result with file position and code
